# node/swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "description":"This is the api for users"
    },
    "tags":[
        {
            "name":"PUBLIC ROUTE",
            "description":"Public Route"
        }
    ],
    "paths":{
        "/":{
            "summary":"Testing Health Check Route",
            "get":{
                "description":"User Health Check",
                "tags":["PUBLIC ROUTE"],
                "responses":{
                    "200":{
                        "description":"User Health Pass"
                    }
                }
            }
        },
        "/location":{
            "summary":"Testing location Check Route",
            "get":{
                "description":"User location Check",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "name":"x-auth-token",
                        "in":"header",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User location Pass"
                    }
                }
            }
        },
        
        "/restaurants":{
            "summary":"Get restaurants",
            "get":{
                "description":"Retrieve restaurants based on state, local government, and kitchen.",
                "tags":["Restaurants"],
                "parameters":[
                    {
                        "name":"x-auth-token",
                        "in":"header",
                        "required":true
                    },
                    {
                        "name":"stateId",
                        "in":"query",
                        "required":true,
                        "description":"The ID of the state",
                        "schema":{
                            "type":"integer"
                        }
                    },
                    {
                        "name":"lgaId",
                        "in":"query",
                        "required":true,
                        "description":"The ID of the local government",
                        "schema":{
                            "type":"integer"
                        }
                    },
                    {
                        "name":"kitchenId",
                        "in":"query",
                        "required":false,
                        "description":"The ID of the kitchen type",
                        "schema":{
                            "type":"integer"
                        }
                    },
                    {
                        "name":"mealTypesId",
                        "in":"query",
                        "required":false,
                        "description":"The ID of the mealTypesId",
                        "schema":{
                            "type":"integer"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"Successful response",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "items":{
                                        "$ref":"#/components/schemas/Restaurant"
                                    }
                                }
                            }
                        }
                    },
                    "400":{
                        "description":"Bad request"
                    },
                    "404":{
                        "description":"Not Found"
                    },
                    "500":{
                        "description":"Internal Server"
                    }
                }
            }
        },
        
        "/mealType":{
            "summary":"Testing mealType Check Route",
            "get":{
                "description":"User mealType Check",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "name":"x-auth-token",
                        "in":"header",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User restaurants Pass"
                    }
                }
            }
        },
        "/filters/{id}":{
            "summary":"Testing location Check Route",
            "get":{
                "description":"User location Check",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "name":"x-auth-token",
                        "in":"header",
                        "required":true
                    },
                    {
                        "name":"id",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User location Pass"
                    }
                }
            }
        },
        "/menus":{
            "summary":"Testing menus Check Route",
            "get":{
                "description":"User menus Check",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "name":"x-auth-token",
                        "in":"header",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User menus Pass"
                    }
                }
            }
        },
        "/menu/{restId}/{lgaId}/{stateId}":{
            "summary":"Testing menu Check Route",
            "get":{
                "description":"User menu Check",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "name":"x-auth-token",
                        "in":"header",
                        "required":true,
                        "schema":{
                            "type":"string"
                        }
                    },
                    {
                        "name":"restId",
                        "in":"path",
                        "required":true,
                        "schema":{
                            "type":"integer"
                        }
                    },
                    {
                        "name":"lgaId",
                        "in":"path",
                        "required":true,
                        "schema":{
                            "type":"integer"
                        }
                    },
                    {
                        "name":"stateId",
                        "in":"path",
                        "required":true,
                        "schema":{
                            "type":"integer"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User menu Pass"
                    }
                }
            }
        },
        "/restDetails/{id}":{
            "summary":"Testing restaurant details Check Route",
            "get":{
                "description":"User restaurants details Check",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "name":"x-auth-token",
                        "in":"header",
                        "required":true
                    },
                    {
                        "name":"id",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User restaurant details Pass"
                    }
                }
            }
        },
        "/menuDetails/{id}":{
            "summary":"Testing menu details Check Route",
            "get":{
                "description":"User menu details Check",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "name":"x-auth-token",
                        "in":"header",
                        "required":true
                    },
                    {
                        "name":"id",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User menu details Pass"
                    }
                }
            }
        },
        "/menuDetails":{
            "summary":"Testing menuDetails Check Route",
            "post":{
                "description":"User menuDetails Check",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                
                                    "id":[1,2,3]
                                
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User menuDetails Pass"
                    }
                }
            }
        },
        "/placeOrder":{
            "summary":"Testing placeOrder Check Route",
            "post":{
                "description":"User placeOrder Check",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                
                                "orderId" : 2, 
                                "name" : "Amit", 
                                "email" : "amit@gmail.com", 
                                "address" : "Hom 65", 
                                "phone" : 8934645457, 
                                "cost" : 612, 
                                "menuItem" : 
                                [ 
                                    45, 34, 41 
                                ], 
                                "status" : "Pending"
                                
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User placeOrder Pass"
                    }
                }
            }
        },
        "/orders":{
            "summary":"Testing orders Check Route",
            "get":{
                "description":"User orders Check",
                "tags":["PUBLIC ROUTE"],
                "responses":{
                    "200":{
                        "description":"User orders Pass"
                    }
                }
            }
        },
        "/updateOrder":{
            "summary":"Testing updateorder Check Route",
            "put":{
                "description":"User updateorders Check",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "_id":"6702d568a8f9b8bb8716a3b4",
                                "status":"On the way"
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User updateorders Pass"
                    }
                }
            }
        },
        "/deleteOrder":{
            "summary":"Testing deleteOrder Check Route",
            "delete":{
                "description":"User deleteOrder Check",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "_id":"6702d52e0b6e1398e56a8681"
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User deleteOrder Pass"
                    }
                }
            }
        }
    },
    "components": {
            "schemas": {
            "Restaurant": {
                "type": "object",
                "properties": {
                "restaurant_id": {
                    "type": "integer",
                    "description": "The unique ID of the restaurant."
                },
                "restaurant_name": {
                    "type": "string",
                    "description": "The name of the restaurant."
                },
                "restaurant_address": {
                    "type": "string",
                    "description": "The address of the restaurant."
                },
                "stateId": {
                    "type": "integer",
                    "description": "The ID of the state the restaurant is located in."
                },
                "lgaId": {
                    "type": "integer",
                    "description": "The ID of the local government area."
                },
                "kitchenId": {
                    "type": "integer",
                    "description": "The ID of the kitchen type."
                }
                }
            }
            }
        }
        }